✨ Sync playlist completed Youtube, Cloud, App (#11)
* 👔 Sync Completed Event Type

* ♻️ Refactor youtube and cloud models

* 👔 Upload youtube Playlist to the Cloud

* 👔 Fetch Cloud Playlists

---------
