+++ ibvn/Modules/Common/List/ListViewModel.swift
@@ -6,43 +6,66 @@
 //
 
 import Foundation
+import FirebaseFirestore
 
-final class ListViewModel: ObservableObject {
-    // MARK: Property Wrappers
-    @Published var youtubePlaylists: YoutubePlaylists = .init()
+final class ListViewModel: ObservableObject, PresentAlertType {
+   // MARK: Property Wrappers
+    @Published var cloudPlaylists: [CloudPlaylist] = []
+    @Published var alertInfo: AlertInfo?
     
     // MARK: Propertes
     let ibvnType: IbvnType
+    
+    var alertIsPresenting: Bool = false
     
     // MARK: Initialization
     init(ibvnType: IbvnType) {
         self.ibvnType = ibvnType
        
-        localFetchElRetoDeHoyPlaylists()
+        fetchCloudPlaylists()
     }
     
     // MARK: Functions
-    func localFetchElRetoDeHoyPlaylists() {
-        guard let url = Bundle.main.url(forResource: ibvnType.localDataFileName, withExtension: "json") else {
-            print("json file not found")
-            
-            return
-        }
+    func fetchCloudPlaylists() {
+        let dataBase = Firestore.firestore()
         
-        guard let data = try? Data(contentsOf: url) else {
-            print("error getting Data from json")
-            
-            return
-        }
-        
-        DispatchQueue.main.async {
-            do {
-                let youtubePlaylists = try JSONDecoder().decode(YoutubePlaylists.self, from: data)
-//                print("🚩 elRetoDeHoyLists: \(String(describing: youtubePlaylists))")
-                self.youtubePlaylists = youtubePlaylists
-            } catch let error as NSError {
-                print("🚩 error decoding local #ElRetoDeHoy: \(String(describing: error))")
+        dataBase.collection("playlists").addSnapshotListener { [weak self] snapshot, error in
+            guard let data = snapshot?.documents, error == nil else {
+                self?.setupAlertInfo(AlertInfo(title: "Firebase Error",
+                                               message: "No se ha logrado recuperar datos.",
+                                               type: .error,
+                                               leftButtonConfiguration: .okConfiguration))
+                
+                return
             }
+            
+            self?.cloudPlaylists = data.enumerated().map({ playlist in
+                return CloudPlaylist(
+                    id: playlist.element["id"] as? String ?? "",
+                    publishedAt: playlist.element["publishedAt"] as? String ?? "",
+                    title: playlist.element["title"] as? String ?? "",
+                    description: playlist.element["description"] as? String ?? "",
+                    thumbnailUrl: playlist.element["thumbnailUrl"] as? String ?? "",
+                    thumbnailWidth: playlist.element["thumbnailWidth"] as? Int ?? 0,
+                    thumbnailHeight: playlist.element["thumbnailHeight"] as? Int ?? 0
+                )
+            })
+            
+            guard
+                self?.ibvnType == .elRestoDeHoy ||
+                self?.ibvnType == .nocheDeViernes
+            else {
+                return
+            }
+            
+            self?.cloudPlaylists = self?.cloudPlaylists.filter({ playlist in
+                playlist.title.contains(self?.ibvnType.hashTag ?? "") ||
+                playlist.description.contains(self?.ibvnType.hashTag ?? "")
+            }) ?? []
         }
+    }
+    
+    func setupAlertInfo(_ alert: AlertInfo) {
+        presentAlert(alert)
     }
 }
