+++ ibvn/Modules/Model/CloudLive.swift
@@ -1,5 +1,5 @@
 //
-//  LiveCloud.swift
+//  CloudLive.swift
 //  ibvn
 //
 //  Created by Jose Letona on 10/4/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct LiveCloud: Encodable {
+struct CloudLive: Encodable {
     let videoId: String
     
     init(videoId: String = "") {
