+++ ibvn/Modules/Model/CloudPlaylist.swift
@@ -1,5 +1,5 @@
 //
-//  PlaylistCloud.swift
+//  CloudPlaylist.swift
 //  ibvn
 //
 //  Created by Jose Letona on 14/4/24.
@@ -8,21 +8,8 @@
 import Foundation
 
 // MARK: - Item
-struct PlaylistCloud: Codable {
+struct CloudPlaylist: Codable {
     let id: String
-    let snippet: SnippetCloud
-    
-    init(
-        id: String = "",
-        snippet: SnippetCloud = .init()
-    ) {
-        self.id = id
-        self.snippet = snippet
-    }
-}
-
-// MARK: - Snippet
-struct SnippetCloud: Codable {
     let publishedAt: String
     let title: String
     let description: String
@@ -30,16 +17,8 @@ struct SnippetCloud: Codable {
     let thumbnailWidth: Int
     let thumbnailHeight: Int
     
-    enum CodingKeys: CodingKey {
-        case publishedAt
-        case title
-        case description
-        case thumbnailUrl
-        case thumbnailWidth
-        case thumbnailHeight
-    }
-    
     init(
+        id: String = "",
         publishedAt: String = "",
         title: String = "",
         description: String = "",
@@ -47,6 +26,7 @@ struct SnippetCloud: Codable {
         thumbnailWidth: Int = 0,
         thumbnailHeight: Int = 0
     ) {
+        self.id = id
         self.publishedAt = publishedAt
         self.title = title
         self.description = description
