+++ ibvn/Modules/Settings/SettingsViewModel.swift
@@ -11,11 +11,11 @@ import Moya
 
 class SettingsViewModel: ObservableObject, PresentAlertType {
     // MARK: Property Wrappers
-    @Published var youtubeLive: Live = .init()
+    @Published var youtubeLive: YoutubeLive = .init()
     @Published var viewMessage: String = ""
     @Published var alertInfo: AlertInfo?
-    @Published var playlistResponse: PlaylistResponse = .init()
-    @Published var playlistCloud: [PlaylistCloud] = .init()
+    @Published var youtubePlaylists: YoutubePlaylists = .init()
+    @Published var cloudPlaylists: [CloudPlaylist] = .init()
     
     // MARK: Properties
     let provider = MoyaProvider<YoutubeApiManager>()
@@ -28,30 +28,29 @@ class SettingsViewModel: ObservableObject, PresentAlertType {
     // MARK: Functions
     func syncPlaylist(pageToken: String = "") {
         viewMessage = pageToken.isEmpty ? "" : viewMessage
-        playlistCloud = pageToken.isEmpty ? .init() : playlistCloud
+        cloudPlaylists = pageToken.isEmpty ? .init() : cloudPlaylists
         
         provider.request(.playlist(pageToken: pageToken)) {[weak self] result in
             switch result {
             case let .success(response):
                 do {
-                    self?.playlistResponse = try JSONDecoder().decode(PlaylistResponse.self, from: response.data)
-                    _ = self?.playlistResponse.items.map { [weak self] playlist in
-                        self?.playlistCloud.append(PlaylistCloud(id: playlist.id,
-                                                                snippet: SnippetCloud(
-                                                                    publishedAt: playlist.snippet.publishedAt,
-                                                                    title: playlist.snippet.title,
-                                                                    description: playlist.snippet.description,
-                                                                    thumbnailUrl: playlist.snippet.thumbnails.medium.url,
-                                                                    thumbnailWidth: playlist.snippet.thumbnails.high.width,
-                                                                    thumbnailHeight: playlist.snippet.thumbnails.high.height
-                                                                )
-                                                               )
+                    self?.youtubePlaylists = try JSONDecoder().decode(YoutubePlaylists.self, from: response.data)
+                    _ = self?.youtubePlaylists.items.map { [weak self] playlist in
+                        self?.cloudPlaylists.append(CloudPlaylist(id: playlist.id,
+                                                                  publishedAt: playlist.snippet.publishedAt,
+                                                                  title: playlist.snippet.title,
+                                                                  description: playlist.snippet.description,
+                                                                  thumbnailUrl: playlist.snippet.thumbnails.medium.url,
+                                                                  thumbnailWidth: playlist.snippet.thumbnails.high.width,
+                                                                  thumbnailHeight: playlist.snippet.thumbnails.high.height)
                         )
                     }
-                    self?.viewMessage += "\n🔄 \(self?.playlistCloud.count ?? 0) Listas descargadas."
+                    self?.viewMessage += "\n🔄 \(self?.cloudPlaylists.count ?? 0) Listas descargadas."
                     
-                    guard let nextPageToken = self?.playlistResponse.nextPageToken, !nextPageToken.isEmpty else {
-                        self?.viewMessage += "\n✅ \(self?.playlistCloud.count ?? 0) Listas descargadas en total."
+                    guard let nextPageToken = self?.youtubePlaylists.nextPageToken, !nextPageToken.isEmpty else {
+                        self?.viewMessage += "\n✅ \(self?.cloudPlaylists.count ?? 0) Listas descargadas en total."
+                        
+                        self?.saveListsOnCloud(cloudPlaylist: self?.cloudPlaylists ?? [])
                         
                         return
                     }
@@ -74,7 +73,7 @@ class SettingsViewModel: ObservableObject, PresentAlertType {
             switch result {
             case let .success(response):
                 do {
-                    self.youtubeLive = try JSONDecoder().decode(Live.self, from: response.data)
+                    self.youtubeLive = try JSONDecoder().decode(YoutubeLive.self, from: response.data)
                     
                     if self.youtubeLive.items.isEmpty {
                         self.viewMessage += "\n 🚫 \(eventType), no disponible."
@@ -103,8 +102,28 @@ class SettingsViewModel: ObservableObject, PresentAlertType {
         }
     }
     
+    private func saveListsOnCloud(cloudPlaylist: [CloudPlaylist]) {
+        viewMessage += "\n⬆️ Subiendo \(cloudPlaylist.count) Listas a la nube."
+        
+        let dataBase = Firestore.firestore()
+        
+        for playlist in cloudPlaylist {
+            dataBase.collection("playlists")
+                .document(playlist.id)
+                .setData(playlist.asDictionary()) { [weak self] error in
+                    guard error == nil else {
+                        self?.displayError(error)
+                        
+                        return
+                    }
+                }
+        }
+        
+        viewMessage += "\n🆗 \(cloudPlaylist.count) Listas en la Nube"
+    }
+    
     private func saveLiveOnCloud(liveVideoId: String) {
-        let liveCloud = LiveCloud(videoId: liveVideoId)
+        let liveCloud = CloudLive(videoId: liveVideoId)
         
         let dataBase = Firestore.firestore()
         
@@ -117,7 +136,8 @@ class SettingsViewModel: ObservableObject, PresentAlertType {
                     return
                 }
                 
-                self?.viewMessage += "\n ✅ En vivo en la nube"
+                self?.viewMessage += "\n 🆗 En vivo en la nube"
             }
+        
     }
 }
